(356) Move fetchWeather service to services/ folder
